+++ RELEASE-NOTES.md
@@ -4,6 +4,28 @@ This is a pre-release version and APIs will change quickly. Before `1.0` release
 
 Please note after `1.0` Semver will be followed using normal protocols.
 
+# Version 0.12.0
+* **Tailwind** - Added a plugin for using Tailwind inside Semantic UI components. This can be used alongside the css tokens provided by the css framework, or instead of it. This will scan your component javascript, html and css and attach only the tailwind styles used.
+
+> Note: the plugin code may be modified if `defineComponent` gets a formal 'Plugin API'. For now you will need to pass in the component to the plugin before defining the component.
+
+
+```javascript
+import { TailwindPlugin } from '@semantic-ui/tailwind';
+const tailwindPlugin = TailwindPlugin();
+
+
+// Transform with Tailwind plugin
+const definition = {
+  // your component definition
+  template: `<div class="p-4 bg-red-500"></div>`
+}
+defineComponent( await tailwindPlugin(definition) );
+
+```
+
+* **Template Compiler** - Fixed an issue with nested conditionals inside svg.
+
 # Version 0.11.3
 * **Input** - Fix placeholder focused color to be theme aware